backend/Auth/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appwrite==9.0.2
asgiref==3.8.1
attrs==24.3.0
bidict==0.23.1
blis==1.2.0
CacheControl==0.14.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cohere==5.13.4
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-allauth==65.5.0
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docstring_parser==0.16
fastavro==1.10.0
filelock==3.16.1
firebase-admin==6.7.0
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
google-api-core==2.24.0
google-api-python-client==2.165.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain==0.3.23
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langchain-community
langchain_pinecone
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
numpy
oauthlib==3.2.2
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parameterized==0.9.0
pillow==11.0.0
pinecone-client
preshed==3.0.9
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==3.11.4
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
sentence-transformers
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.3.1
srsly==2.5.1
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.4
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.8.2
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
